fix: constructor should throw if it fails to initialize
