+++ mediapipe_api/framework/calculator_graph.cc
@@ -24,15 +24,6 @@ MpReturnCode mp_CalculatorGraph__(mediapipe::CalculatorGraph** graph_out) {
 
 void mp_CalculatorGraph__delete(mediapipe::CalculatorGraph* graph) { delete graph; }
 
-MpReturnCode mp_CalculatorGraph__PKc(const char* text_format_config, mediapipe::CalculatorGraph** graph_out) {
-  TRY_ALL
-    mediapipe::CalculatorGraphConfig config;
-    auto result = google::protobuf::TextFormat::ParseFromString(text_format_config, &config);
-    *graph_out = result ? new mediapipe::CalculatorGraph(config) : nullptr;
-    RETURN_CODE(MpReturnCode::Success);
-  CATCH_ALL
-}
-
 MpReturnCode mp_CalculatorGraph__PKc_i(const char* serialized_config, int size, mediapipe::CalculatorGraph** graph_out) {
   TRY_ALL
     auto config = ParseFromStringAsCalculatorGraphConfig(serialized_config, size);
