+++ mediapipe_api/framework/calculator_graph.h
@@ -29,7 +29,6 @@ typedef std::map<std::string, mediapipe::Packet> SidePackets;
 typedef absl::Status* NativePacketCallback(mediapipe::CalculatorGraph* graph, int stream_id, const mediapipe::Packet&);
 
 MP_CAPI(MpReturnCode) mp_CalculatorGraph__(mediapipe::CalculatorGraph** graph_out);
-MP_CAPI(MpReturnCode) mp_CalculatorGraph__PKc(const char* text_format_config, mediapipe::CalculatorGraph** graph_out);
 MP_CAPI(MpReturnCode) mp_CalculatorGraph__PKc_i(const char* serialized_config, int size, mediapipe::CalculatorGraph** graph_out);
 MP_CAPI(void) mp_CalculatorGraph__delete(mediapipe::CalculatorGraph* graph);
